I have revised all functions accordingly, dividing them into separate folders and departments.
