+++ Malshinon.csproj
@@ -102,8 +102,8 @@
     <Compile Include="Models\People.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\IntelSubmissionService.cs" />
     <Compile Include="Services\ReporterAnalysisService.cs" />
+    <Compile Include="Services\IntelSubmissionService.cs" />
     <Compile Include="Dal\IntelReportDal.cs" />
   </ItemGroup>
   <ItemGroup>
